src/requirements.txt
azure-storage-blob==0.37.1